feat: Enhance category management with validation and improved UI elements

+++ app/Http/Controllers/CategoriesController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers;
 
 use App\Models\Categories;
@@ -31,7 +30,7 @@ class CategoriesController extends Controller
      */
     public function store(CategoriesRequest $request)
     {
-       
+
         $category = Categories::create($request->only('title', 'description', 'status'));
 
         if ($request->hasFile('image')) {
@@ -55,23 +54,29 @@ class CategoriesController extends Controller
     public function edit($id)
     {
         $categories = Categories::findOrFail($id);
+
         return view('categories.edit', compact('categories'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateCategoryRequest $request,$id)
+    public function update(UpdateCategoryRequest $request, $id)
     {
         $categories = Categories::findOrFail($id);
-        
+
+        //check inactive if products are associated with this category
+        if ($categories->products()->exists() && $request->status == 'inactive') {
+            return redirect()->route('categories.index')->with('error', 'Category cannot be inactive as it has associated products.');
+        }
+
         // Check if the image is present in the request
         if ($request->hasFile('image')) {
             $categories->deleteImage();
             $categories->addMedia($request->file('image'))->toMediaCollection('image');
         }
         $categories->update($request->only('title', 'description', 'status'));
-       
+
         return redirect()->route('categories.index')->with('success', 'Category updated successfully.');
     }
 
@@ -80,6 +85,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Categories $category)
     {
+
+        //check inactive if products are associated with this category
+        if ($category->products()->exists()) {
+            return redirect()->route('categories.index')->with('error', 'Category cannot be inactive as it has associated products.');
+        }
+
         // Check if the category has any associated media
         if ($category->hasMedia('image')) {
             $category->deleteImage();
